--- app/src/main/java/com/example/lawjoin/BoardFreeActivity.kt
@@ -1,28 +1,0 @@
-package com.example.lawjoin
-
-import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Adapter
-import android.widget.Button
-import android.widget.EditText
-import androidx.appcompat.app.AppCompatActivity
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.example.lawjoin.data.model.Post
-import com.google.firebase.firestore.FirebaseFirestore
-
-class BoardFreeActivity: AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_free_board_list)
-
-    }
-
-}
-
-
-
