+++ app/src/main/java/com/example/lawjoin/data/model/Notice.kt
