+++ app/src/main/java/com/example/lawjoin/post/BoardFreeActivity.kt
@@ -1,0 +1,17 @@
+package com.example.lawjoin.post
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.lawjoin.R
+
+class BoardFreeActivity: AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_free_board_list)
+
+    }
+
+}
+
+
+
