+++ app/src/main/java/com/example/lawjoin/post/PostAdapter.kt
@@ -1,10 +1,11 @@
-package com.example.lawjoin
+package com.example.lawjoin.post
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.example.lawjoin.R
 import com.example.lawjoin.data.model.Post
 
 class PostAdapter(private val postData: List<Post>) : RecyclerView.Adapter<PostAdapter.PostViewHolder>(){
